+++ CHANGES.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- BREAKING: added `RoomServiceClient` wrapper class
+    - Wrapper around the generated twirp `RoomServiceClient`
+    - Has nicer API which is more similar to the official JavaScript LiveKit Server SDK.
+    - This is a breaking change because generated `RoomServiceClient` used to be exposed directly.
+
 ### Changed
 
 - build: add check for isort command
@@ -30,6 +37,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - [LiveKit protocol v0.11.0](https://github.com/livekit/protocol/releases/tag/v0.11.0)
 
 [Unreleased]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.1.1...HEAD
+
+[0.2.0]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.2.0...v0.1.1
 
 [0.1.1]: https://github.com/tradablebits/livekit-server-sdk-python/compare/v0.1.0...v0.1.1